+++ src/main/java/com/nubble/backend/post/comment/service/CommentService.java
@@ -4,6 +4,7 @@ import com.nubble.backend.post.comment.domain.Comment;
 import com.nubble.backend.post.comment.service.CommentCommand.CommentCreateCommand;
 import com.nubble.backend.post.comment.service.CommentCommand.CommentDeleteCommand;
 import com.nubble.backend.post.comment.service.factory.CommentFactory;
+import com.nubble.backend.post.comment.service.remover.CommentRemover;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
